Initial DWA planner commit
This commit adds the dynamic window approach (DWA) planner classes and
a ROS node to interface with it (based on the PF node). The CMake
lists for both ROS and ROS2 have been changed to build the additional
node.

An example Python launch file compatible with ROS 2 is included. The
parameters in the launch file do not match a specific vehicle and may
require some tweaking to work in a specific scenario.

The planner takes advantage of the speed control node, which bypasses
the pure pursuit algorithm for lateral control while maintaining PID
control in the longitudinal degree of freedom. This commit fixes two
small bugs in the speed control node relating to parsing the throttle
coefficient from the ROS parameter server and clamping the Z component
of the published twist message.