Fix DWA planner bugs
This commit fixes a number of bugs in the DWA planner. Specifically, it
corrects a typo which caused obstacle distances to be computed
incorrectly and a typo which caused the global path deviation term to
always use the last computed trajectory as reference. Finally, the speed
cost term is tweaked to correctly handle AGV velocities higher than
those used for planning.

Credit to @stefanwapnick for spotting the issues with the obstacles and
global path deviation cost terms.