Add MPC planner Julia C wrapper node
This commit adds a node that wraps the MPC planner from
TulgaErsal/AVT-341-MPC.

The commit adds the node as an optional build target in the ROS 2
CMakeLists. The target is only built if the correct version of Julia is
available on PATH. The commit also adds a platform-independent mechanism
for CMake to detect a Julia installation.

The commit includes a short documentation with the steps required to
set up the Julia C API. The steps currently focus on deployment on a
Linux operating system.

In its current state, the node launches the Julia C API and registers
subscription callbacks and publishers, interfacing with a mock Julia
module with a minimal API.