Add preliminary segmentation grid support
This commit adds support for the segmentation grid to the DWA planner.
The segmentation grid is parsed separately from the occupancy grid under
the assumption that the two share the same geometry. The segmentation
grid is accessed only at the locations occupied by the predicted
trajectory and its cost is included in the objective function as the
cumulative segmentation cost of all states in the current trajectory
scaled by a user-defined weight. The ROS node also exposes an additional
parameter `dwa_thresh_seg`: values strictly higher than this specified
threshold in the segmentation grid will be considered untraversable and
yield an artificially high cost in the objective function.. The
parameter defaults to  a value of `100`, such that no value in the
segmentation grid triggers this behaviour on default parameters.

This commit also removes all cost normalisation in the planner. All
terms in the cost function are now to be considered in their
SI units and their values scaled by the user-specified cost to yield
comparable values for a given scenario. The planner users will have to
tune their DWA planner parameters from scratch after this commit.

This commit also removes the ability to choose a different motion model
than Ackermann in the planner, as the synchro drive robots of the
original DWA formulation are not target platforms for the AVT-341
autonomy stack.