Add external modules and sysimage generation
This commit allows the inclusion of external modules for the Julia MPC
parameters and motion models through either CMake variables or ROS
parameters. Moreover, the commit includes a small and tested template
for generating a Julia sysimage at CMake compile time.

This commit also adds function handles to run the Julia MPC planner and
set the desired heading, as well as including temporary versions of the
Julia scripts summoned by the wrapper.