Add Sphinx documentation
This commit adds documentation pages for the repository. The
documentation is built using Sphinx and its extension Breathe for
automatically generating the C++ API pages.

A Python requirements file is included to help with local builds of the
documentation. Moreover, the Makefile of the documentation is adapted to
automatically run Doxygen and generate an up-to-date C++ API before the
build process.

The installation procedure for ROS and ROS 2 has been moved from the
main README file to the documentation pages, and instructions on how to
build the documentation (on Ubuntu 22.04) have been added to the README
file.

A part of the NATURE stack documentation, publicly available on GitHub,
has been adapted to reStructuredText and included. Moreover, the
Markdown files included in the repository root have been removed and
included directly in the documentation as reStructuredText.

This feature is not yet complete: parts of the NATURE stack
documentation have yet to be ported and the C++ API currently only
includes the DWA planner.