Add occupancy grid parser node
This commit adds a new node to parse cells from an occupancy grid into a
set of discrete obstacles. The node includes options to filter the
obstacles based on AGV odometry information and user preference.

In addition to the above changes, to complement the interface of the new
node, the commit includes new message definitions for generic static
obstacles of different shape and size, as a unit or as a collection.

Lastly, the commit includes brief documentation pages for the node and a
simple occupancy grid test node publishing a mock occupancy grid message
from an image loaded from disk or from a predefined map to test the node
functionality.