Add camera detection messages
This commit adds messages for serializing the output of a
two-dimensional (image frame) object detector to the stack interfaces.

The messages are close in specification to the ROS `vision_msgs`, with
only small modifications to the definitions. In future commits, it
should be possible to align the two interfaces such that either one may
be chosen by the user.