Add Docker container for the ROS Noetic stack
This commit adds a Dockerfile and a matching compose file to build a
Docker container for the ROS Noetic version of the autonomy stack.

At the current time, the container only supports basic functionality of
the stack (as there is no GPU support, no MATLAB runtime installation
and no Julia installation).

In addition to the Compose file, a development container configuration
file is included for easy deployment using Dev Containers in supported
development environments.