Fix compiler warning in ROS 2 build
This commit fixes the compiler warnings raised in the ROS 2 build of the
stack due to the inclusion of an outdated header from the
`tf2_geometry_msgs` library.

Moreover, the commit removes a compile definition to suppress warnings
regarding the use of the deprecated PointCloud message raised by the
`sensor_msgs` package, as this option no longer has any effect and the
warnings are shown regardless.