Refactor CMakeLists file for the ROS 2 build
This commit refactors the root CMakeLists file for the ROS 2 build of
the stack to improve clarity and make future additions easier.

The changes split the "global" install target for all nodes into
individual targets to allow for conditional installs in the future.
Moreover, the node proxy is now compiled as a library and linked against
each node target, and all indirect "ament" dependencies are dropped from
node targets as a result.

Finally, the commit enforces more consistent spacing and separates
"runtime" from "test" nodes.