Fix ROS dev container mount point
This commit fixes the mount point in the ROS Noetic dev container to
point to the repository root. Moreover, it adds sourcing of the ROS
environments directly in the container build, as in the interactive
shell of dev container the Bash configuration file is executed normally.

Finally, the commit adds the `git` package to the container build to
allow for git operations from inside the dev container.