Add Poetry dependency manager option
This commit adds the configuration files required to create a Poetry
environment to build the project documentation. The gitignore file is
update to accomodate the Poetry environment lock file and the README
instructions are updated to included the Poetry build as an option.

In addition to the changes above, the commit adds the link to the
generated documentation in PDF format to the README file and moves the
pip requirements file to the root of the repository.