Restructure documentation contents
This commit restructures the sections of the documentation to improve
the look and feel of the PDF build.

The commit introduces the todo items extension for Sphinx and marks
incomplete sections with a todo item. All child nodes in the
documentation source are now consistently renamed to "index" and located
each under its own subfolder. Download links internal to the
documentation build have been explicitly linked (through raw LaTeX) to a
Dropbox shared folder in the PDF build.

Finally, the commit fixes the rendering of the API index page by
enforcing a single column and avoiding text overlap.