Refactor build system
This commit introduces extensive refactoring of the build scripts for
both ROS versions of the stack, following the "one CMakeLists file per
build target" approach.

The ROS 1 and ROS 2 CMakeLists files are split into individual modules
through the `add_subdirectory()` CMake command to improve readability
and editability of the build files. To accomodate for the restrictions
of the `add_subdirectory()` command, additional `ros1` and `ros2`
folders are added accordingly for each build target, and common source
files are grouped under a parent directory.

Some nodes remain to be adapted to this new approach (e.g. UAB
perception node, AI nodes and testing nodes). Some modules still refer
to each other sources and a common library remains to be added for these
nodes (e.g. communications and mission management nodes).

The commit also fixes the ROS 1 build of the stack, which was previously
not linking the node proxy library at runtime correctly.