Add detection model training node
This commit adds a node to train custom detection models in YOLOv8 to
the perception module of the stack.

As this node is the first fully-integrated Python node of the stack, the
commit includes some boilerplate to create a parent Python package for
all autonomy stack features written in Python.

The commit also includes preliminary documentation on the object
detection node and instructions on how to use the provided Poetry
environment.