Add ROS 2 Docker container and devcontainer
This commit adds two Dockerfile to build the ROS 2 version of the
stack with and without CUDA support. The build recipes are
complemented by compose files to launch individual nodes and
Visual Studio Code devcontainer configuration files to develop
directly within the application.

Image are tagged as either "devel", suitable for development or
devcontainers, or "runtime", including a pre-built version of the
stack and suitable for testing or production.