Add LiDAR-only tracking
This commit is a large rewrite of the functionality of the object
tracking node to allow for tracking the lead vehicle using a previous
camera detection and the current point cloud from a LiDAR sensor when
new detections are missing.

The commit also includes fixes to the transforms between camera frame
and world frame to ensure consistency across different setups.

Finally, the commit adds dynamic parameter reconfiguration for all
parameters accessed during the node spin.